+++ BDFrameWork-UnityFull/Assets/Code/Game/ScreenView/ScreenView_Demo2_Windows.cs
@@ -19,8 +19,8 @@ public class ScreenView_Demo2_Windows : IScreenView
 		 //一定要设置为true，否则当前是未加载状态
 		 this.IsLoad = true;
 		 //打开 window 2
-		 UIMgr.Inst.LoadWindows((int) WinEnum.Win_Test2);
-		 UIMgr.Inst.ShowWindow((int)WinEnum.Win_Test2);
+		 UIMgr.Inst.LoadWindows(WinEnum.Win_Test2);
+		 UIMgr.Inst.ShowWindow(WinEnum.Win_Test2);
 		 //
 		 Debug.Log("进入demo2");
 	 }
